feat(collateral-contract): get collateral account info using address of an authorized signer
adds a function that takes in a signer and receiver. Then matches the signer to the sender they are authorized to sign for. Finally, return the corresponding collateral account info for the  sender-receiver pair. This should allow a receiver to check available collateral based on a recovered address from a signature without specifically knowing the address of the sender.

Signed-off-by: Bryan Cole <bryan.cole@semiotic.ai>